+++ Mapwize/Mapwize/Classes/MWZGeometry.h
@@ -1,6 +1,7 @@
 #import <Foundation/Foundation.h>
 #import <CoreLocation/CoreLocation.h>
 #import "MWZCoordinate.h"
+#import "MWZBounds.h"
 
 @interface MWZGeometry : NSObject
 
@@ -11,5 +12,7 @@
 - (instancetype) initWithDictionary:(NSDictionary*) dictionary;
 
 - (NSDictionary*) toDictionary;
+
+- (MWZBounds*) getBounds;
 
 @end