+++ Mapwize/Mapwize/Classes/MWZGeometryPolygon.h
@@ -13,6 +13,8 @@
 
 - (NSArray*) getCoordinates;
 
+- (MWZBounds*) getBounds;
+
 - (NSDictionary*) toDictionary;
 
 @end