+++ Mapwize/Mapwize/Classes/MWZTranslation.h
@@ -3,7 +3,7 @@
 @interface MWZTranslation : NSObject
 
 @property(nonatomic, strong) NSString* title;
-@property(nonatomic, strong) NSString* subtitle;
+@property(nonatomic, strong) NSString* subTitle;
 @property(nonatomic, strong) NSString* details;
 @property(nonatomic, strong) NSString* language;