+++ Mapwize/Mapwize/Classes/MWZVenue.h
@@ -19,4 +19,6 @@
 
 - (instancetype)initFromDictionary:(NSDictionary*)dic;
 
+- (MWZBounds*) getBounds;
+
 @end